Use int instead of int8_t for the 3 variables
Since strcmp return a int, and since the spec
of strcmp do not tell the return value, it
could return 256 and this would overflow.

Found by Coverity scan.
(thanks to Stéphane Marcheusin who explained
the details to me)

BUG: 789278
Reviewed-on: https://review.gluster.org/16738
Smoke: Gluster Build System <jenkins@build.gluster.org>
Tested-by: Michael Scherer <misc@fedoraproject.org>
NetBSD-regression: NetBSD Build System <jenkins@build.gluster.org>
CentOS-regression: Gluster Build System <jenkins@build.gluster.org>

+++ rpc/rpc-lib/src/rpc-transport.c
@@ -171,7 +171,7 @@ rpc_transport_load (glusterfs_ctx_t *ctx, dict_t *options, char *trans_name)
 	char *type = NULL;
 	char str[] = "ERROR";
 	int32_t ret = -1;
-	int8_t is_tcp = 0, is_unix = 0, is_ibsdp = 0;
+	int is_tcp = 0, is_unix = 0, is_ibsdp = 0;
 	volume_opt_list_t *vol_opt = NULL;
         gf_boolean_t bind_insecure = _gf_false;
         xlator_t   *this = NULL;
